+++ src/main/java/io/github/astrapi69/swing/panels/lottery/LotteryPanel.java
@@ -36,7 +36,6 @@ import javax.swing.*;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.Model;
 import io.github.astrapi69.swing.base.BasePanel;
-import io.github.astrapi69.swing.icon.StringIcon;
 
 public class LotteryPanel extends BasePanel<LotteryBox> implements ActionListener
 {